ProblemSolver stores ConstraintsAndComplements_t
